Cast token idx to int64

+++ bittensor/utils/tokenizer_utils.py
@@ -1044,7 +1044,7 @@ def topk_tokens_to_vocab_size(topk_tensor: torch.Tensor, vocab_size_std: int, vo
     batch_size, topk_p1, max_len = topk_tensor.shape  # [batch_size, (topk + 1), max_len]
     topk = topk_p1 - 1
 
-    topk_tokens = topk_tensor[:, :-1, 1].round().int()  # [batch_size, topk] first tokens
+    topk_tokens = topk_tensor[:, :-1, 1].round().to(torch.int64)  # [batch_size, topk] first tokens
     topk_probs = topk_tensor[:, :-1, 0]  # [batch_size, topk] Probabilities for each phrase in topk
     floor_probs = topk_tensor[:, -1, 0]  # [batch_size] Floor probabilities as mean probability for non-topk tokens
